Dynamically load unmanaged ICU library
This allows the library to work with multiple ICU versions. This makes
it easier to produce a nuget package that works with any ICU version
(rather than having to create a separate package for each ICU version),
and it allows to use the same binary/nuget package for different
Linux versions which come pre-installed with different ICU versions.