Refactor two methods
This change removes two native methods that we had in two versions
although they mapped to the same unmanaged method. At least one
of them caused problems when running tests in LibPalaso because it
mapped an unmanaged object to an IntPtr instead of the SafeHandle.
This caused the pointer to get GC'ed with Mono 3.x, causing the test
to fail. This change replaces the call of this overload with the
save version. The other method wasn't used at all.