Don't update AssemblyVersion for new patch
`AssemblyVersion` only changes when major or minor version number
changes (instead of `AssemblyFileVersion` accidentally introduced
in previous patch version). This is necessary so that the assembly
signature doesn't change and icu.net.dll referenced in a project
can be replaced with a bugfix version without requiring to change
the binding redirect.