Allow to specify preferred directory for unmanaged ICU
This change adds the `Wrapper.SetPreferredIcu4cDirectory()` method that
allows to specify a directory where we look for the unmanaged ICU library
first. This is particular useful if you want to confine a specific
version.

This fixes the problem where it wasn't possible to use an ICU version
not located in or below the current assembly directory if the current
directory contained a different icu4c version.

This change also adds `runtimes/win7-*/native` as additional location
to search for the ICU binaries.

+semver:minor