In OneTimeSetup limit version only on Windows
On Windows we install Icu4c from nuget packages, and so we know
which versions are available and can limit to that. On Linux
we use whatever icu4c version we find on the system and skip
several tests that require icu4c installed from nuget. This
change modifies confining the ICU version in the `SetupFixture`
so that it runs only on Windows.