Reference System.ValueTuple 4.5 in Framework and 4.4 in .NET Standard
This is a modification to a recent changeset, which simply downgraded the version to 4.4 for all targets. This change preserves original functionality for Framework targets and only changes the .NET Standard. In some cases, consuming the .NET Standard package was causing an ugly warning, which moving to 4.4 fixes. Although it's not confirmed that this would break things on Framework, it's safer to keep it using the other version.

+semver:minor